bokeh works, gonna use webcam for video processing
