+++ src/pages/StudentWebcam.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 
 import { currentUser } from "./LoginForm";
 import EndExam from "../components/alerts/EndExam";
@@ -33,6 +33,12 @@ import {
 } from "livekit-client";
 import patchData from "../hooks/patchData";
 import useUsers from "../hooks/useUsers";
+import * as bodySegmentation from "@tensorflow-models/body-segmentation";
+import "@tensorflow/tfjs-core";
+// Register WebGL backend.
+import "@tensorflow/tfjs-backend-webgl";
+import * as tf from "@tensorflow/tfjs";
+import "@mediapipe/selfie_segmentation";
 
 let name = "";
 
@@ -50,6 +56,8 @@ const StudentWebcam = () => {
   const [room, setRoom] = useState<Room | null>(null);
   const { data, loading, error } = useUsers();
   const localVideoRef = useRef(null);
+  const canvasRef = useRef<HTMLCanvasElement>(null);
+
   const [ready, isReady] = useState<boolean>(false);
   const [lkParticipant, setLkParticipant] = useState<any>(null);
   const [isConnected, setIsConnected] = useState(false);
@@ -117,6 +125,92 @@ const StudentWebcam = () => {
     connectToRoom();
   }, [token]);
 
+  const applyBokehEffect = useCallback(async () => {
+    const video = localVideoRef.current;
+    const canvas = canvasRef.current;
+    if (!video || !canvas) return;
+
+    const segmenter = await bodySegmentation.createSegmenter(
+      bodySegmentation.SupportedModels.MediaPipeSelfieSegmentation,
+      {
+        runtime: "mediapipe",
+        solutionPath:
+          "https://cdn.jsdelivr.net/npm/@mediapipe/selfie_segmentation",
+        modelType: "general",
+      }
+    );
+
+    const foregroundThreshold = 0.5;
+    const backgroundBlurAmount = 20;
+    const edgeBlurAmount = 5;
+    const flipHorizontal = false;
+
+    const drawEffect = async () => {
+      if (!video || video.readyState !== 4) return;
+
+      if (video.videoWidth > 0 && video.videoHeight > 0) {
+        canvas.width = video.videoWidth;
+        canvas.height = video.videoHeight;
+        video.width = video.videoWidth;
+        video.height = video.videoHeight;
+      } else {
+        console.error("Video dimensions not available");
+      }
+
+      const segmentation = await segmenter.segmentPeople(video);
+
+      if (!segmentation) {
+        console.error("Segmentation failed!");
+        return;
+      }
+
+      if (canvas.width === 0 || canvas.height === 0) {
+        console.error("Canvas dimensions are zero!");
+        return;
+      }
+
+      const ctx = canvas.getContext("2d");
+      if (!ctx) {
+        console.error("Failed to get canvas rendering context!");
+        return;
+      }
+
+      await bodySegmentation.drawBokehEffect(
+        canvas,
+        video,
+        segmentation,
+        foregroundThreshold,
+        backgroundBlurAmount,
+        edgeBlurAmount,
+        flipHorizontal
+      );
+
+      requestAnimationFrame(drawEffect);
+    };
+
+    drawEffect();
+  }, [localVideoRef, canvasRef]);
+
+  useEffect(() => {
+    const video = localVideoRef.current;
+    if (!video) return;
+
+    const handleVideoPlay = () => {
+      if (video.videoWidth === 0 || video.videoHeight === 0) {
+        console.error("Video dimensions are zero!");
+        return;
+      }
+      applyBokehEffect();
+    };
+
+    video.addEventListener("play", handleVideoPlay);
+
+    return () => {
+      // Cleanup event listener on component unmount.
+      video.removeEventListener("play", handleVideoPlay);
+    };
+  }, [localVideoRef, applyBokehEffect]);
+
   const publishTracks = async (participant: LocalParticipant) => {
     await participant.setCameraEnabled(true);
     await participant.setMicrophoneEnabled(false);
@@ -152,7 +246,6 @@ const StudentWebcam = () => {
     name = currentUser.name;
   }
 
-  const canvasRef = useRef<HTMLCanvasElement>(null);
   const [frameCaptureInterval, setFrameCaptureInterval] = useState<
     number | null
   >(null);
@@ -310,19 +403,35 @@ const StudentWebcam = () => {
           justifyContent="center"
           alignItems="center"
         >
-          <video
-            style={{ width: "50%", borderRadius: "10px", overflow: "hidden" }}
-            ref={localVideoRef}
-            autoPlay={true}
-            muted={true}
-          />
+          <div style={{ position: "relative" }}>
+            <canvas
+              ref={canvasRef}
+              style={{
+                borderRadius: "10px",
+                overflow: "hidden",
+                width: "50%",
+                height: "auto",
+              }}
+            ></canvas>
+            <video
+              style={{
+                width: "50%",
+                // borderRadius: "10px",
+                // overflow: "hidden",
+                visibility: "hidden",
+              }}
+              ref={localVideoRef}
+              autoPlay={true}
+              muted={true}
+            />
+          </div>
         </Box>
-        <canvas
+        {/* <canvas
           ref={canvasRef}
           width={640}
           height={480}
           style={{ display: "none" }}
-        ></canvas>{" "}
+        ></canvas>{" "} */}
         {/* Hide the canvas element */}
         <div hidden={ready ? true : false}>
           <p>This is where the checklist will be</p>
